requirements-dev.txt
# Development Dependencies
-r requirements.txt

# Testing
pytest>=7.4.2
pytest-flask>=1.2.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
factory-boy>=3.2.0
faker>=19.0.0

# Code Quality
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.3.0

# Security
safety>=2.3.0
bandit>=1.7.0

# Documentation
Sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0

# Development Tools
ipython>=8.0.0
jupyter>=1.0.0
notebook>=6.5.0

# Debugging
flask-debugtoolbar>=0.13.0
pdbpp>=0.10.0

# Performance
memory-profiler>=0.60.0
line-profiler>=4.0.0

# Database
alembic>=1.11.0
sqlalchemy-utils>=0.41.0

# API Testing
requests>=2.31.0
httpx>=0.24.0

# Monitoring
sentry-sdk[flask]>=1.30.0
